backend/requirements.txt
anthropic>=0.30.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
requests>=2.32.0
python-multipart>=0.0.9
pydantic>=2.0.0
